refactor: add new subpackages core, utils and entities
