+++ src/diffhouse/core/__init__.py
@@ -1,0 +1,15 @@
+"""Git metadata extractors."""
+
+from .branches import get_branches
+from .changed_files import stream_changed_files
+from .commits import stream_commits
+from .diffs import stream_diffs
+from .tags import get_tags
+
+__all__ = [
+    'stream_commits',
+    'stream_changed_files',
+    'stream_diffs',
+    'get_branches',
+    'get_tags',
+]
