+++ src/diffhouse/core/changed_files.py
@@ -1,60 +1,12 @@
 import re
 from collections.abc import Iterator
 from contextlib import contextmanager
-from dataclasses import asdict, dataclass
 from io import StringIO
 
+from ..entities import ChangedFile
 from ..git import GitCLI
 from .constants import RECORD_SEPARATOR
 from .utils import fast_hash_64, safe_iter, split_stream
-
-
-@dataclass(slots=True, frozen=True)
-class ChangedFile:
-    """Snapshot of a file that was modified in a specific commit."""
-
-    def to_dict(self) -> dict:
-        """Convert the object to a dictionary.
-
-        Returns:
-            A dictionary representation of the changed file.
-
-        """
-        return asdict(self)
-
-    commit_hash: str
-    """Full hash of the commit."""
-    path_a: str
-    """Path to the file before applying the commit's changes."""
-    path_b: str
-    """
-    Path to the file after applying the commit's changes.
-
-    Differs from `path_a` for renames and copies.
-    """
-    changed_file_id: str
-    """
-    Unique record identifier hashed from `commit_hash`, `path_a`, and `path_b`.
-    """
-    change_type: str
-    """
-    Single-letter code representing the change type.
-
-    Most commonly one of `A` (added), `C` (copied), `D` (deleted), `M`
-    (modified) or `R` (renamed). See
-    [git-status](https://git-scm.com/docs/git-status#_short_format) for all
-    possible values.
-    """
-    similarity: int
-    """
-    Similarity index between the two file versions.
-
-    `0`-`100` for renames and copies, `100` otherwise.
-    """
-    lines_added: int
-    """Number of lines added to the file in the commit."""
-    lines_deleted: int
-    """Number of lines deleted from the file in the commit."""
 
 
 def stream_changed_files(path: str) -> Iterator[ChangedFile]:
