+++ src/diffhouse/core/constants.py
