--- src/diffhouse/engine/__init__.py
@@ -1,18 +1,0 @@
-"""Git command parsers and data structures."""
-
-from .branches import get_branches
-from .changed_files import ChangedFile, stream_changed_files
-from .commits import Commit, stream_commits
-from .diffs import Diff, stream_diffs
-from .tags import get_tags
-
-__all__ = [
-    'Commit',
-    'stream_commits',
-    'ChangedFile',
-    'stream_changed_files',
-    'Diff',
-    'stream_diffs',
-    'get_branches',
-    'get_tags',
-]
