+++ src/diffhouse/entities/__init__.py
@@ -1,0 +1,7 @@
+"""Git metadata entities."""
+
+from .changed_file import ChangedFile
+from .commit import Commit
+from .diff import Diff
+
+__all__ = ['Commit', 'ChangedFile', 'Diff']
